Script Update
I’ve updated the script to make it automatically make a backup of your
currently installed npm formula/packages as a text file inside the
directory you are cd’ed into at the time of execution.

This should help people reinstall formula/packages if they later decide
to install Node again, or via Homebrew or MacPorts instead of the
website .pkg, or vice versa and so on.