Updates
Temporary move from /bin/bash to /usr/local/bin/bash. This was largely
done because this script executes the use of sudo, and the OS X
provided /bin/bash has already proven exploitable to privilege
escalation (See CVE-2014-6271). This is a temporary move, and will be
reverted as soon as Apple release a fix to the heavily vulnerable
provided Bash. (Yes, Homebrew’s bash is still partially vulnerable, but
less vulnerable is still better than nothing).